Add de-duplication of ITS_LIVE STAC and
* refactor hyp3 specific stuff to a hyp3.py module
* refactor S1 specific stuff to the sentinel1.py module
